+++ app/src/main/java/me/magnum/melonds/impl/emulator/AndroidEmulatorManager.kt
@@ -153,7 +153,7 @@ class AndroidEmulatorManager(
         return MelonEmulator.getRewindWindow()
     }
 
-    override fun getFps(): Int {
+    override fun getFps(): Float {
         return MelonEmulator.getFPS()
     }
